+++ R/parse.r
@@ -20,6 +20,8 @@ parse_post_list <- function(post_list) {
                                    .default = NA_integer_),
     like_count    = purrr::map_int(post_list, "likeCount",
                                    .default = NA_integer_),
+    quote_count    = purrr::map_int(post_list, "quoteCount",
+                                    .default = NA_integer_),
     indexed_at    = parse_time(purrr::map_chr(post_list, "indexedAt")),
     # TODO: return URL instead of URI
     in_reply_to   = purrr::map_chr(post_list, c("record", "reply", "parent", "uri"),
@@ -171,21 +173,41 @@ parse_likes <- function(res) {
 }
 
 parse_actors <- function(res) {
-  tibble::tibble(
-    did               = purrr::map_chr(res, "did"),
-    indexed_at        = parse_time(purrr::map_chr(res, "indexedAt",
-                                                  .default = NA_character_)),
-    actor_handle      = purrr::map_chr(res, "handle"),
-    actor_name        = purrr::map_chr(res, "displayName",
-                                       .default = NA_character_),
-    actor_description = purrr::map_chr(res, "description",
-                                       .default = NA_character_),
-    actor_avatar      = purrr::map_chr(res, "avatar",
-                                       .default = NA_character_),
-    viewer_data       = purrr::map(res, "viewer",
-                                   .default = NA_character_),
-    labels_data       = purrr::map(res, "labels",
-                                   .default = NA_character_)
-  )
+  purrr::map(res, as_tibble_onerow) |>
+    dplyr::bind_rows() |>
+    dplyr::rename(actor_handle = "handle",
+                  actor_name = "display_name",
+                  actor_description = "description",
+                  actor_avatar  = "avatar")
 }
 
+parse_starter_packs <- function(res) {
+
+  all <- purrr::pluck(res, "starterPack")
+  users <- purrr::pluck(all, "listItemsSample")  |>
+    purrr::map(purrr::list_flatten) |>
+    purrr::map(as_tibble_onerow) |>
+    dplyr::bind_rows()
+  colnames(users) <- stringr::str_remove(colnames(users), "subject_")
+  colnames(users) <- paste0("user_", colnames(users))
+  all$listItemsSample  <- NULL
+  list_info <- all |>
+    purrr::list_flatten() |>
+    purrr::list_flatten() |>
+    as_tibble_onerow()
+  colnames(list_info) <- paste0("list_", colnames(list_info))
+  dplyr::bind_cols(users, list_info)
+}
+
+parse_list <- function(res, resp) {
+
+  users <- res  |>
+    purrr::map(purrr::list_flatten) |>
+    purrr::map(as_tibble_onerow) |>
+    dplyr::bind_rows()
+  list_info <- purrr::pluck(resp, "list") |>
+    purrr::list_flatten() |>
+    as_tibble_onerow()
+  colnames(list_info) <- paste0("list_", colnames(list_info))
+  dplyr::bind_cols(users, list_info)
+}