+++ NEWS.md
@@ -1,9 +1,21 @@
+# atrrr 0.0.5
+
+* adds three new functions for lists: get_list(), get_list_feed(), and get_starter_pack()
+* adapted to httr2 1.1.0
+* now supports posting videos,several images and compresses images that are too large
+* updated auto functions to reflect changes to the endpoints
+* added new parameters to search_post()
+* improves stability of preview cards in posts
+* improves parsing of actor data
+
+
 # atrrr 0.0.4
 
 * adds option to include language, additional tags and labels to post
 * fix issue with preview cards for cases with multiple links
 * add search function to vignette (thank you for noticing, @SoaresAlisson)
 * address issue with new version of glue package (thank you, @jennybc)
+* add ability to post videos
 
 # atrrr 0.0.3